xkbcomp: Fix the int type of ExprInteger
Avoid implicit conversion from `int64_t`.

+++ src/xkbcomp/ast-build.c
@@ -44,7 +44,7 @@ ExprCreateString(xkb_atom_t str)
 }
 
 ExprDef *
-ExprCreateInteger(int ival)
+ExprCreateInteger(int64_t ival)
 {
     ExprDef *expr = ExprCreate(EXPR_VALUE, EXPR_TYPE_INT, sizeof(ExprInteger));
     if (!expr)
